LayoutTests/fast/dom/Element/scroll-width-visible-expected.txt

Tests that element.clientWidth of visible element with scrollbars returns proper values

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById('scrollDiv').clientWidth is 190
PASS successfullyParsed is true

TEST COMPLETE
